+++ client/src/app/auth/pages/login-empresas/page.jsx
@@ -1,6 +1,6 @@
 import AuthHeader from "@/app/auth/AuthHeader";
 import LoginFormEmpresa from "./components/LoginForm";
-import { Card } from "@/components/ui/Card";
+import { Card } from "@/app/recovery/components/ui/Card";
 
 export default function LoginEmpresaPage() {
   return (
